requirements.txt
asgiref==3.5.2; python_version >= '3.7'
django==4.0.6
django-appconf==1.0.5; python_version >= '3.1'
django-crispy-forms==1.14.0
django-imagekit==4.1.0
django-jalali==6.0.0
django-jalali-date==1.0.1
jdatetime==4.1.0; python_version >= '3.7'
pillow==9.2.0
sqlparse==0.4.2; python_version >= '3.5'
